token-metrics-service/requirements.txt
bidict
blinker
certifi
charset-normalizer
click
eventlet
Flask
flask-cors
Flask-SocketIO
Flask-SQLAlchemy
greenlet
h11
idna
itsdangerous
Jinja2
MarkupSafe
python-dotenv
python-engineio
python-socketio
redis
requests 
simple-websocket
SQLAlchemy 
typing_extensions 
urllib3 
Werkzeug
wsproto
